Статус фильма + Пользовательский рейтинг + Рефактор
1. Реализовал WatchFilmCollectionLinksCubit для прослушивания статуса фильма (сохранён ли в локальное хранилище). Подключил к UI на экранах SearchFilmScreen и SearchedFilmsScreen
2. Реализовал RatingBottomSheet и логику для обновления userRating на экране FilmInformation. Добавил UpdateRatingFilmInformationUseCase.
3. Рефактор SearchedFilmsScreen и SearchedFilmsBloc
4. Ренейминг классов

+++ lib/data/models/film_base_model.dart
@@ -18,7 +18,7 @@ class FilmBaseModel {
   @HiveField(10) String? posterUrl;
   @HiveField(11) String? posterUrlPreview;
   @HiveField(12) String? userComment;
-  @HiveField(13) double? userRating;
+  @HiveField(13) int? userRating;
 
   FilmBaseModel({
     this.kinopoiskId,
